+++ src/relstorage/adapters/mysql/locker.py
@@ -36,7 +36,7 @@ _SET_TIMEOUT_STMT = 'SET SESSION innodb_lock_wait_timeout = %s'
 _SET_TIMEOUT_DEFAULT_STMT = _SET_TIMEOUT_STMT % ('DEFAULT',)
 
 @contextmanager
-def lock_timeout(cursor, timeout):
+def lock_timeout(cursor, timeout, restore_to=None):
     """
     ContextManager that sets the lock timeout to the given value,
     and returns it to the DEFAULT when done.
@@ -44,11 +44,17 @@ def lock_timeout(cursor, timeout):
     If *timeout* is ``None``, makes no changes to the connection.
     """
     if timeout is not None: # 0 is valid
+        # Min value of timeout is 1; a value less than that produces
+        # a warning but gets truncated to 1
+        timeout = timeout if timeout >= 1 else 1
         cursor.execute(_SET_TIMEOUT_STMT, (timeout,))
         try:
             yield
         finally:
-            cursor.execute(_SET_TIMEOUT_DEFAULT_STMT)
+            if restore_to is None:
+                cursor.execute(_SET_TIMEOUT_DEFAULT_STMT)
+            else:
+                cursor.execute(_SET_TIMEOUT_STMT, (restore_to,))
     else:
         yield
 
@@ -95,6 +101,10 @@ class MySQLLocker(AbstractLocker):
     advisory locks anyway.
     """
 
+    # The old MySQL 5.7 syntax is the default
+    _lock_share_clause = 'LOCK IN SHARE MODE'
+    _lock_share_clause_nowait = 'LOCK IN SHARE MODE'
+
     def __init__(self, options, driver, batcher_factory):
         super(MySQLLocker, self).__init__(options, driver, batcher_factory)
         self._supports_row_lock_nowait = None
@@ -123,15 +133,20 @@ class MySQLLocker(AbstractLocker):
             ver = self._metadata_to_native_str(ver)
             major = int(ver[0])
             __traceback_info__ = ver, major
-            self._supports_row_lock_nowait = (major >= 8)
+            native_nowait = self._supports_row_lock_nowait = (major >= 8)
+            if native_nowait:
+                self._lock_share_clause = 'FOR SHARE'
+                self._lock_share_clause_nowait = 'FOR SHARE NOWAIT'
+            else:
+                assert self._lock_readCurrent_oids_for_share
+                self._lock_readCurrent_oids_for_share = self.__lock_readCurrent_nowait
 
     def _on_store_opened_set_row_lock_timeout(self, cursor, restart=False):
-        if restart:
-            return
         self._set_row_lock_timeout(cursor, self.commit_lock_timeout)
 
     def _set_row_lock_timeout(self, cursor, timeout):
-        # Min value of timeout is 1
+        # Min value of timeout is 1; a value less than that produces
+        # a warning.
         timeout = timeout if timeout >= 1 else 1
         cursor.execute(self.set_timeout_stmt, (timeout,))
         # It's INCREDIBLY important to fetch a row after we execute the SET statement;
@@ -140,10 +155,19 @@ class MySQLLocker(AbstractLocker):
         # corrupted data, written after free?' or something like that.
         cursor.fetchone()
 
+
+    def __lock_readCurrent_nowait(self, cursor, current_oids):
+        # For MySQL 5.7, we emulate NOWAIT by setting the lock timeout
+        if self.lock_readCurrent_for_share_blocks:
+            return AbstractLocker._lock_readCurrent_oids_for_share(self, cursor, current_oids)
+
+        with lock_timeout(cursor, 0, self.commit_lock_timeout):
+            return AbstractLocker._lock_readCurrent_oids_for_share(self, cursor, current_oids)
+
     def release_commit_lock(self, cursor):
         "Auto-released by transaction end."
 
-    def _get_commit_lock_debug_info(self, cursor):
+    def _get_commit_lock_debug_info(self, cursor, was_failure=False):
         cursor.execute('SELECT connection_id()')
         conn_id = str(cursor.fetchone()[0])
 
@@ -153,8 +177,12 @@ class MySQLLocker(AbstractLocker):
             return 'Connection: ' + conn_id + '\n' + self._rows_as_pretty_string(cursor)
         except self.driver.driver_module.Error:
             # MySQL 5, or no permissions
-            cursor.execute('SELECT * from information_schema.innodb_locks')
-            rows = self._rows_as_pretty_string(cursor)
+            try:
+                cursor.execute('SELECT * from information_schema.innodb_locks')
+                rows = self._rows_as_pretty_string(cursor)
+            except self.driver.driver_module.Error:
+                # MySQL 8, and we had no permissions.
+                return 'Connection: ' + conn_id
             return 'Connection: ' + conn_id + '\n' + rows
 
     def hold_pack_lock(self, cursor):