+++ src/relstorage/adapters/postgresql/locker.py
@@ -30,28 +30,35 @@ class PostgreSQLLocker(AbstractLocker):
     def _on_store_opened_set_row_lock_timeout(self, cursor, restart=False):
         # This only lasts beyond the current transaction if it
         # commits.
-        self._set_row_lock_timeout(cursor, self.commit_lock_timeout)
+        if not restart:
+            # We never change our lock timeout setting; the lock_and_choose_tid
+            # functions have it as part of their function definition, and
+            # we have NOWAIT for when we need a 0 timeout.
+            self._set_row_lock_timeout(cursor, self.commit_lock_timeout)
 
     def _set_row_lock_timeout(self, cursor, timeout):
         # This will rollback if the transaction rolls back,
         # which should restart our store.
         # Maybe for timeout == 0, we should do SET LOCAL, which
         # is guaranteed not to persist?
-        cursor.execute('SET lock_timeout = %s', (timeout,))
+        # Recall postgresql uses milliseconds, but our configuration is given
+        # in seconds.
+        self.driver.set_lock_timeout(cursor, timeout * 1000)
 
     def release_commit_lock(self, cursor):
         # no action needed, locks released with transaction.
         pass
 
-    def _get_commit_lock_debug_info(self, cursor):
+    def _get_commit_lock_debug_info(self, cursor, was_failure=False):
         # When we're called, the transaction is guaranteed to be
         # aborted, so to be able to execute any queries we need to
         # rollback. Unfortunately, that would release the locks that
-        # we're holding (which may already have been released?
-        # Unclear.)
-        #
-        # This method isn't specifically tested under those failures
-        # circumstances, and probably doesn't work.
+        # we're holding, which have probably(?) already have been released
+        # anyway. The output can be excessive.
+        if was_failure:
+            cursor.connection.rollback()
+            return 'Transaction failed; no lock info available'
+
         from . import debug_locks
         debug_locks(cursor)
         return self._rows_as_pretty_string(cursor)